Fixed pasting issue when vim is run within tmux.
If run within a tmux session, vim no longer uses the unnamed register(")
for yank/paste. This means y and p now works as expected. A side-effect
is that pasting from the system clipboard on MacOS no longer works as
expected. A wrapper/workaround is needed, but left out at this time as
OS X is not the only platform this files are used on.