Add restart command for ssh-agent.
Theres now an alias for restarting the ssh-agent and prompting for the
password when theres an issue. Handy for the occasional "refused
connection" issue, sometimes experienced after a session has been
restored, with tmux running inside of it.

Also added some clarifying comments to the other aliases defined in the
file.