Fixed vim bindings for inserting new lines.
Previously there were two bindings defined in .vimrc, allowing the user
to insert a new line before or after the cursor, without entering insert
mode. These were bound to C-o and C-O, respectively. This was a mistake,
as the terminal receives both of those key combinations as the same,
with the letter capitalised. Therefore it is not possible to bind both
non capital and capital letters in combination with CTRL.

Instead, C-O is now bound to inserting a line after the cursor, exiting
insert mode and moving one line up, effectively simulating inserting a
line after the cursor without altering its position or the editor's
mode. This binding works regardless of the shift button being pressed.

C-P is a new binding, allowing the user to do the same as above but
inserting a line above the cursor, instead. This binding was bound to
moving the cursor one line up, previously.