Added vale for prose linting.
Vale is now the default tool used to lint prose, using A.L.E. It's used
for .txt and .md. If it's the only supported tool installed it will
automatically be used for any filetype that's associated with it.

Note: Vale is syntax-aware which means it will lint files including that
have special syntax such as markdown, without generating errors for the
syntax itself.