Fixed bug in .tmux.conf
When running .tmux_snapshot.conf (file used to initialise the tmux
status bar) as part of .tmux.conf, on startup of tmux, it would fail.

This was because the PATH used for this command was not relational and
assumed the file would be present in the current working directory. It
now looks at the home directory always.